Reorganized files to make them easier to update
Created: data_files dir for files that are used by the server code but
don't need to be updated in the source control.

build_tools dir to hold the tools we need for building, but don't need
to be updated on the server.

Created config_files.php to hold filenames and folder paths that are
shared by all server files (so it's easier to manage files that are
accessed by multiple files).

Added the ConditionId to the transition data object.